+++ addons/pt/saft/tax_combo.go
@@ -48,7 +48,7 @@ func normalizeTaxCombo(tc *tax.Combo) {
 		case tax.KeyOutsideScope:
 			tc.Ext = tc.Ext.
 				Set(ExtKeyTaxRate, TaxRateExempt).
-				Set(ExtKeyExemption, "M99")
+				SetOneOf(ExtKeyExemption, "M99", "M44", "M45", "M46")
 		case tax.KeyIntraCommunity:
 			tc.Ext = tc.Ext.
 				Set(ExtKeyTaxRate, TaxRateExempt).
@@ -82,7 +82,7 @@ func prepareTaxComboKey(tc *tax.Combo) {
 		tc.Key = tax.KeyExport
 	case "M16":
 		tc.Key = tax.KeyIntraCommunity
-	case "M99":
+	case "M99", "M44", "M45", "M46":
 		tc.Key = tax.KeyOutsideScope
 	case "M01", "M02", "M03", "M06", "M07", "M09", "M10", "M11",
 		"M12", "M13", "M14", "M15", "M19", "M20", "M21", "M25",